// BikeShop.Frontend/webclient/src/entities/models/NovaPoshta/Response/NPCityResponse.ts
export interface NPCityResponse {
    Description: string;
    DescriptionRu: string;
    Ref: string;
    Delivery1: string;
    Delivery2: string;
    Delivery3: string;
    Delivery4: string;
    Delivery5: string;
    Delivery6: string;
    Delivery7: string;
    Area: string;
    AreaDescription: string;
    AreaDescriptionRu: string;
    SettlementType: string;
    IsBranch: string;
    PreventEntryNewStreetsUser: string | null;
    Conglomerates: string | null;
    CityID: string;
    SettlementTypeDescriptionRu: string;
    SettlementTypeDescription: string;
}